# api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: URL Shortener API
  version: 1.0.0
  description: A simple and efficient URL shortening service API
  contact:
    name: URL Shortener Team
    url: https://github.com/drementer/url-shortener
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/roles/{id}:
    get:
      summary: Get Role by ID
      description: Retrieve details of a specific role by its ID. Requires ADMIN role.
      tags: [Roles]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The role ID
      responses:
        '200':
          description: Role details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Role not found

    patch:
      summary: Update Role
      description: Update the name, description, or maxActiveLinks quota of an existing role. Requires ADMIN role.
      tags: [Roles]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The role ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          description: Validation error, including an empty request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: At least one field must be provided
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Role not found
        '409':
          description: Role name conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete Role
      description: >-
        Delete a role. The built in ADMIN role cannot be deleted, and neither
        can a role that is still assigned to at least one user. Requires ADMIN
        role.
      tags: [Roles]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The role ID
      responses:
        '204':
          description: Role deleted successfully, no content returned
        '400':
          description: The built in ADMIN role cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Cannot delete the built-in ADMIN role
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Role not found
        '409':
          description: The role is still assigned to at least one user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Role is still assigned to 3 user(s)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        The access token from a register, login or refresh response, sent as
        `Authorization: Bearer <accessToken>`. It expires after 15 minutes by
        default, after which the refresh token buys a new one.
  responses:
    Unauthorized:
      description: The access token is missing, malformed or expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing:
              summary: No token sent
              value:
                error: Authentication required
            invalid:
              summary: Token does not verify
              value:
                error: Invalid or expired token
    Forbidden:
      description: The caller lacks permissions or quota to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            insufficientPermissions:
              summary: Insufficient permissions
              value:
                error: Insufficient permissions
            quotaExceeded:
              summary: Active link quota exceeded
              value:
                error: Active link quota exceeded (maximum 10 active links allowed for role EDITOR)
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Error message
          example: URL not found
        details:
          type: array
          description: >-
            Present on validation failures only. Carries every issue the schema
            found, so a client can place each message on the field it belongs
            to. The top level error repeats the first of these.
          items:
            type: object
            required: [path, message]
            properties:
              path:
                type: string
                description: Dotted path of the offending field
                example: customSlug
              message:
                type: string
                example: This custom slug is reserved
...
    Role:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          description: Unique role identifier (UUID)
          example: a91b7c2d-4e3f-4a5b-8c6d-7e8f9a0b1c2d
        name:
          type: string
          description: Unique role name
          example: EDITOR
        description:
          type: string
          nullable: true
          description: Description of the role and tier
          example: Editor user with expanded link quota
        maxActiveLinks:
          type: integer
          nullable: true
          description: Maximum number of concurrently active links (null represents unlimited)
          example: 10
...
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 30
          description: Updated role name
          example: VIP_PRO
        description:
          type: string
          nullable: true
          maxLength: 255
          description: Updated role description
          example: Updated VIP tier description
        maxActiveLinks:
          type: integer
          nullable: true
          minimum: 0
          description: Updated active links quota
          example: 100
...